server/requirements.txt
openenv-core[core]>=0.2.2
fastapi
uvicorn
pydantic>=2.0
openai
